Created ExtractSentenceTexts
Created this new project in RosieTools and got it to work.  It takes all
the files in John's test-sentences directory and for each one creates a
new file which just contains the bare sentence texts instead of the Soar
rules.  The changed file names end with "-new.script" instead of
".soar".